♻️ refactor(tokens): move token-related functions to dedicated module
- relocate functions from utils.py to tokens.py for better organization
- improve maintainability by grouping related functionalities
